Rename `Ids` -> `IdSyntax`.
A bit more descriptive / less ambiguous.
