+++ local-modules/@bayou/config-common-default/tests/test_IdSyntax.js
@@ -5,7 +5,7 @@
 import { assert } from 'chai';
 import { describe, it } from 'mocha';
 
-import { Ids } from '@bayou/config-common-default';
+import { IdSyntax } from '@bayou/config-common-default';
 
 /** {array<*>} Array of non-strings. */
 const NON_STRINGS = [
@@ -22,51 +22,51 @@ const NON_STRINGS = [
   new Map()
 ];
 
-describe('@bayou/config-common-default/Ids', () => {
+describe('@bayou/config-common-default/IdSyntax', () => {
   describe('isAuthorId()', () => {
     it('should accept 32-character alphanum ASCII strings', () => {
-      assert.isTrue(Ids.isAuthorId('123abc7890ABC456789012'));
+      assert.isTrue(IdSyntax.isAuthorId('123abc7890ABC456789012'));
     });
 
     it('should allow underscores and hyphens', () => {
-      assert.isTrue(Ids.isAuthorId('123456789_123456789-12'));
+      assert.isTrue(IdSyntax.isAuthorId('123456789_123456789-12'));
     });
 
     it('should not allow non-ASCII characters', () => {
-      assert.isFalse(Ids.isAuthorId('123456789•123456789•12'));
+      assert.isFalse(IdSyntax.isAuthorId('123456789•123456789•12'));
     });
 
     it('should not allow non-alphanum characters', () => {
-      assert.isFalse(Ids.isAuthorId('123456789\t123456789+12'));
+      assert.isFalse(IdSyntax.isAuthorId('123456789\t123456789+12'));
     });
 
     it('should throw an error given a non-string argument', () => {
       for (const id of NON_STRINGS) {
-        assert.throws(() => Ids.isAuthorId(id), /badValue/, id);
+        assert.throws(() => IdSyntax.isAuthorId(id), /badValue/, id);
       }
     });
   });
 
   describe('isDocumentId()', () => {
     it('should accept 32-character alphanum ASCII strings', () => {
-      assert.isTrue(Ids.isDocumentId('123abc7890ABC456789012'));
+      assert.isTrue(IdSyntax.isDocumentId('123abc7890ABC456789012'));
     });
 
     it('should allow underscores and hyphens', () => {
-      assert.isTrue(Ids.isDocumentId('123456789_123456789-12'));
+      assert.isTrue(IdSyntax.isDocumentId('123456789_123456789-12'));
     });
 
     it('should not allow non-ASCII characters', () => {
-      assert.isFalse(Ids.isDocumentId('123456789•123456789•12'));
+      assert.isFalse(IdSyntax.isDocumentId('123456789•123456789•12'));
     });
 
     it('should not allow non-alphanum characters', () => {
-      assert.isFalse(Ids.isDocumentId('123456789\t123456789+12'));
+      assert.isFalse(IdSyntax.isDocumentId('123456789\t123456789+12'));
     });
 
     it('should throw an error given a non-string argument', () => {
       for (const id of NON_STRINGS) {
-        assert.throws(() => Ids.isDocumentId(id), /badValue/, id);
+        assert.throws(() => IdSyntax.isDocumentId(id), /badValue/, id);
       }
     });
   });
