+++ local-modules/@bayou/config-common/index.js
@@ -2,10 +2,10 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
-import Ids from './Ids';
+import IdSyntax from './IdSyntax';
 import Text from './Text';
 
 export {
-  Ids,
+  IdSyntax,
   Text
 };
