+++ local-modules/@bayou/doc-common/tests/test_DocumentId.js
@@ -29,10 +29,10 @@ describe('@bayou/doc-common/DocumentId', () => {
     }
   });
 
-  it('should call through to `config-common.Ids.isDocumentId()` to validate strings', () => {
+  it('should call through to `config-common.IdSyntax.isDocumentId()` to validate strings', () => {
     // **TODO:** The check for acceptance depends on the configuration point
-    // `config-common.Ids`. The best we can do here is mock that out and make
-    // sure it's called, but as yet we don't have a standard way to do that
+    // `config-common.IdSyntax`. The best we can do here is mock that out and
+    // make sure it's called, but as yet we don't have a standard way to do that
     // sort of thing, so we're punting for now.
     assert.isTrue(true);
   });
